model/requirements.txt
fastapi 
uvicorn 
onnx
onnxruntime 
python-multipart
pillow 